+++ models/teams.js
@@ -1,6 +1,12 @@
+import { db } from "./db-config.js";
+import { createUserWithEmailAndPassword, getAuth, signInWithEmailAndPassword } from "firebase/auth";
+import { addDoc, collection, doc, getFirestore, setDoc, getDocs, getDoc, updateDoc, deleteDoc, documentId, query as queryFirestore, where } from 'firebase/firestore';
+import { GeoPoint } from "firebase/firestore";
+
+
 // TEAMS
 
-class Team {
+export default class Team {
     constructor(teamData) {
         this._members = teamData.members;
         this._name = teamData.name;
@@ -54,7 +60,7 @@ class Team {
 
 }
 
-function teamGenerator() {
+export function teamGenerator() {
 
   const roles = ['Onsite Operationer', 'Rescue Agent', 'Disaster Operationer', 'Firefighter', 'Police', 'Medic Rescue'];
   const teamNames = ['Quest', 'Giants', 'Squad', 'Alpha', 'Peak'];
@@ -70,7 +76,7 @@ function teamGenerator() {
           leader: false, // Initially set all members as non-leaders
           name: `Member ${i + 1}`,
           role: getRandomRole(),
-          startedAt: new Date().toISOString(),
+          startedAt: new Date(),
       });
   }
 
@@ -78,72 +84,97 @@ function teamGenerator() {
   const leaderIndex = getRandomNumber(0, teamSize - 1);
   members[leaderIndex].leader = true;
 
-  const team = new Team({
+  const team = {
       name: getRandomTeamName(),
       members
-  });
+  };
 
   return team;
 }
 
 // Example usage
-let myTeam = teamGenerator("Alpha Squad");
-console.log(myTeam);
+// let myTeam = teamGenerator("Alpha Squad");
+// console.log(myTeam);
 
 
     // WRITE
-    async function Insert_team(teamData) {
+    export async function InsertTeam(teamData) {
         try {
           const docRef = await addDoc(collection(db, "teams"), teamData);
           console.log("Team added with ID: ", docRef.id);
+          return docRef.id;
         } catch (error) {
-          console.error("Error adding team: ", error);
+          console.error("Error adding team: ", error.stack);
+          return '';
         }
       }
     // WRITE
     
     // READ
-    async function Get_teamById(teamId) {
+    export async function GetTeamById(teamId) {
         try {
           const docRef = doc(db, "teams", teamId);
           const docSnap = await getDoc(docRef);
           if (docSnap.exists()) {
-            console.log("Team data:", docSnap.data());
             return docSnap.data();
           } else {
             console.log("No such team found!");
             return null;
           }
         } catch (error) {
-          console.error("Error fetching team: ", error);
+          console.error("Error fetching team: ", error.stack);
           return null;
         }
     }
     // READ
 
+    async function checkIfTeamExists(teamId) {
+      const teamRef = doc(db, "teams", teamId);
+      try {
+        const teamSnap = await getDoc(teamRef);
+        return teamSnap.exists();
+      } catch (error) {
+        console.error("Error checking team existence: ", error);
+        throw error; // Or handle the error as needed for your application logic
+      }
+    }
+
     // UPDATE
-    async function Update_team(teamId, updatedData) {
+    export async function UpdateTeam(teamId, updatedData) {
         try {
+          if(! await checkIfTeamExists(teamId)) {
+            console.log(`Team does not exist`);
+            return false;
+          }
           const teamRef = doc(db, "teams", teamId);
           await updateDoc(teamRef, updatedData);
           console.log("Team successfully updated");
+          return true;
         } catch (error) {
-          console.error("Error updating team: ", error);
+          console.error("Error updating team: ", error.stack);
+          return false;
         }
     }
     // UPDATE
 
     // DELETE
-    async function Delete_team(teamId) {
+    export async function DeleteTeam(teamId) {
         try {
+          if(! await checkIfTeamExists(teamId)) {
+            console.log(`Team does not exist`);
+            return false;
+          }
           await deleteDoc(doc(db, "teams", teamId));
           console.log("Team successfully deleted");
+          return true;
         } catch (error) {
-          console.error("Error deleting team: ", error);
+          console.error("Error deleting team: ", error.stack);
+          return false;
         }
     }
     // DELETE
 
+/*
 
 let teamObj = {
     members: [
@@ -162,5 +193,7 @@ let teams = {
     skdfjslkds: teamObj,
     dsjhfsdfsd: teamObj
 }
+
+*/
 
 // TEAMS
